Edit dtype of clustered Kaggle to enable bit flipping

+++ mia/datasets.py
@@ -94,7 +94,7 @@ def _prepare_clustered_kaggle(numberOfClusters: int):
     features, _ = _read_kaggle_data()
     kaggleSize = 197324
     labels: NDArray = kmeans.fit_predict(features).reshape(kaggleSize, 1)
-    labels = to_categorical(labels)
+    labels = to_categorical(labels, dtype='int64')
     return Dataset.from_tensor_slices((features, labels))
